Feat : DTO Fix
Former-commit-id: 44341ae78cc1c611f391d92b31a2c8c59582a59c

+++ backend/src/main/java/com/a503/churros/dto/scrap/ScrapInputDTO.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class ScrapInputDTO {
     private String folderName;
 
-    private Integer folderIdx;
+    private Long folderIdx;
 
-    private Integer articleIdx;
+    private Long articleIdx;
 }
